VEGA-834: Добавить data-test-id для текстов валидации (#126)
* feat(test-id): добавление test-id для ошибок

Jira issue: VEGA-834

* test(test-id): фикс test-id

Jira issue: VEGA-834

* fix(test-id): поправил testId для yearStart

Jira issue: VEGA-834

* fix(test-id): исправил нейминг

Jira issue: VEGA-834

* fix(textfield): вынес в отдельный компонент

Jira issue: VEGA-834
