+++ src/ui/core/TextField/TextField.css
@@ -1,0 +1,5 @@
+.VegaTextField {
+  &__ErrorText {
+    margin-top: var(--space-2xs);
+  }
+}
